requirements.txt
Flask
python-dotenv
requests
flask-cors
Flask-Login==0.5.0
Flask-SQLAlchemy==2.1
urllib3==1.26.6
psycopg2-binary==2.9.1
stripe